refactor registry logic into base class
